Use glob patterns for clean command

+++ daves_dev_tools/clean.py
@@ -2,147 +2,209 @@
 This module cleans up files which are ignored by git
 """
 import argparse
-import functools
 import os
 import shutil
+from collections import deque
+from glob import iglob
 from pipes import quote
 from itertools import chain
-from typing import Any, Callable, Dict, FrozenSet, Iterable, Sequence, Set
+from typing import (
+    Dict,
+    Tuple,
+    Iterable,
+    Sequence,
+    Set,
+    FrozenSet,
+    List,
+)
 from .utilities import run
 
 ROOT_DIRECTORY: str = "."
-EXCLUDE_DIRECTORIES: FrozenSet[str] = frozenset(
-    (
-        "./.idea",  # Jetbrains' IDE project settings (Pycharm, Intellij IDEA)
-        "./.vscode",  # Microsoft Visual Studio Code project settings
-        "./.git",  # Git history
-        "./venv",  # Commonly used location for virtual environments
-    )
+DEFAULT_EXCLUDE: Tuple[str, ...] = (
+    ".idea",  # Jetbrains' IDE project settings (Pycharm, Intellij IDEA)
+    ".vscode",  # Microsoft Visual Studio Code project settings
+    ".git",  # Git history
+    "venv",  # Commonly used location for virtual environments
 )
 
-lru_cache: Callable[..., Any] = functools.lru_cache
 
-
-@lru_cache()
-def _is_excluded(
-    absolute_path: str, exclude_directories: FrozenSet[str] = frozenset()
-) -> bool:
-    absolute_path = absolute_path.rstrip("/")
-    for directory_path in exclude_directories:
-        if absolute_path.startswith(f"{directory_path}/"):
-            return True
-    return False
-
-
-@lru_cache()
-def _absolute_sub_directories(
-    root_directory: str, directories: FrozenSet[str]
+def _get_directory_globs_files(
+    directory: str, patterns: Iterable[str], *, recursive: bool = False
 ) -> FrozenSet[str]:
-    directory: str
-    return frozenset(
-        (
-            os.path.abspath(os.path.join(root_directory, directory))
-            for directory in directories
-        )
-    )
+    """
+    Return a `frozenset` of file names matching the glob `pattern` within
+    `directory`, or in a directory matching the glob pattern.
+
+    Parameters:
+
+    - directory (str)
+    - patterns ([str])
+    - recursive (bool) = False
+    """
+    if isinstance(patterns, str):
+        patterns = (patterns,)
+    file_relative_paths: Set[str]
+    relative_paths: FrozenSet[str]
+    current_directory: str = os.path.curdir
+    os.chdir(directory)
+    try:
+        relative_paths = frozenset()
+
+        def add_glob(pattern: str) -> None:
+            nonlocal relative_paths
+            relative_paths |= frozenset(iglob(pattern, recursive=recursive))
+
+        deque(map(add_glob, patterns), maxlen=0)
+        file_relative_paths = set()
+        relative_path: str
+        for relative_path in relative_paths:
+            if os.path.isdir(relative_path):
+                sub_directory: str
+                _: Iterable[str]
+                files: Iterable[str]
+                for sub_directory, _, files in os.walk(relative_path):
+                    sub_directory = sub_directory.replace("\\", "/")
+                    file: str
+                    for file in files:
+                        file_relative_paths.add(f"{sub_directory}/{file}")
+            else:
+                file_relative_paths.add(relative_path.replace("\\", "/"))
+    finally:
+        os.chdir(current_directory)
+    return frozenset(file_relative_paths)
+
+
+def _get_directory_globs(
+    directory: str, patterns: Iterable[str], *, recursive: bool = False
+) -> FrozenSet[str]:
+    """
+    Return a `frozenset` of directory paths matching the glob `pattern` within
+    `directory`, or in a directory matching the glob pattern.
+
+    Parameters:
+
+    - directory (str)
+    - patterns ([str])
+    - recursive (bool) = False
+    """
+    if isinstance(patterns, str):
+        patterns = (patterns,)
+    relative_paths: FrozenSet[str]
+    current_directory: str = os.path.curdir
+    os.chdir(directory)
+    try:
+        relative_paths = frozenset()
+
+        def add_glob(pattern: str) -> None:
+            nonlocal relative_paths
+            relative_paths |= frozenset(iglob(pattern, recursive=recursive))
+
+        deque(map(add_glob, patterns), maxlen=0)
+    finally:
+        os.chdir(current_directory)
+    return relative_paths
 
 
 def get_ignored_files(
     directory: str = ".",
-    exclude_directories: FrozenSet[str] = frozenset(),
-) -> Iterable[str]:
+    exclude: FrozenSet[str] = frozenset(),
+) -> Set[str]:
     """
-    Yield the absolute path of all ignored files, excluding those in any of the
-    `exclude_directories`.
+    Get a `set` containing the relative paths of all ignored files
+    in `directory` excluding those matching any of the glob patterns
+    in `exclude`.
 
     Parameters:
 
-    - root_directory (str): The root project directory.
-    - exclude_directories ({str}) = frozenset(): A `frozenset` of
-      sub-directories to exclude.
+    - directory (str): The root project directory.
+    - exclude ({str}) = frozenset(): A `frozenset` of glob patterns for
+      files and sub-directories to exclude.
     """
-    path: str
-    sub_directory_name: str
-    directories_files: Dict[str, Set[str]] = {}
-    paths: Set[str]
-    directory = os.path.abspath(directory)
-    exclude_directories = _absolute_sub_directories(
-        directory, exclude_directories
-    )
-    path_prefix: str = f'{directory.rstrip("/")}/'
-    path_prefix_length: int = len(path_prefix)
-    quoted_root_dir: str = quote(directory)
-    for path in run(
-        f"git add {quoted_root_dir} && git ls-files -o {quoted_root_dir}",
-        echo=False,
-    ).split("\n"):
-        path = os.path.abspath(f"./{path}")
-        if not _is_excluded(path, exclude_directories):
-            sub_directory_name = ""
-            if "/" in path[path_prefix_length:]:
-                sub_directory_name = path[path_prefix_length:].split("/")[0]
-            if sub_directory_name not in directories_files:
-                directories_files[sub_directory_name] = set()
-            directories_files[sub_directory_name].add(path)
-    for sub_directory_name, paths in directories_files.items():
-        number_of_files: int = len(paths)
-        print(
-            f"Found {number_of_files} ignored "
-            f'file{"s" if number_of_files > 1 else ""} in ./'
-            f"{sub_directory_name}"
+    quoted_directory: str = quote(os.path.abspath(directory))
+    return (
+        set(
+            run(
+                (
+                    f"git add {quoted_directory} && "
+                    f"git ls-files -o {quoted_directory}"
+                ),
+                echo=False,
+            ).split("\n")
         )
-        for path in paths:
-            yield path
+        - _get_directory_globs_files(directory, exclude, recursive=True)
+    )
+
+
+def _is_sub_directory_excluded(
+    sub_directory: str, directory: str, exclude: FrozenSet[str]
+) -> bool:
+    relative_sub_directory: str = os.path.relpath(
+        os.path.abspath(sub_directory), directory
+    )
+    if relative_sub_directory.startswith("./"):
+        relative_sub_directory = relative_sub_directory[2:]
+    relative_sub_directory_list: List[str] = relative_sub_directory.split("/")
+    index: int
+    for index in range(1, len(relative_sub_directory_list)):
+        ancestor: str = "/".join(relative_sub_directory_list[:index])
+        if ancestor in exclude:
+            return True
+    return False
 
 
 def delete_empty_directories(
-    root_directory: str = ".",
-    exclude_directories: FrozenSet[str] = frozenset(),
-    _recurrence: bool = True,
+    directory: str = ".",
+    exclude: FrozenSet[str] = frozenset(),
+    dry_run: bool = False,
+    _recurrence: bool = False,
 ) -> int:
     """
     Deletes empty directories under the current directory.
 
     Parameters:
 
-    - exclude_directories ({str}) = frozenset():
+    - exclude ({str}) = frozenset():
       A set of top-level directory names to exclude
     """
     number_of_deleted_directories: int = 0
-    directory: str
-    directories: Sequence[str]
+    sub_directory: str
+    sub_directories: Sequence[str]
     files: Sequence[str]
-    if not _recurrence:
-        root_directory = os.path.abspath(root_directory)
-        exclude_directories = _absolute_sub_directories(
-            root_directory, exclude_directories
-        )
-    for directory, directories, files in os.walk(
-        root_directory, topdown=False
+    exclude = _get_directory_globs(directory, exclude)
+    for sub_directory, sub_directories, files in os.walk(
+        directory, topdown=False
     ):
-        directory = os.path.abspath(directory)
-        if not _is_excluded(directory, exclude_directories):
-            if not any(
+        if not (
+            _is_sub_directory_excluded(sub_directory, directory, exclude)
+            or any(
                 filter(
-                    lambda name: name != ".DS_Store", chain(directories, files)
+                    lambda name: name != ".DS_Store",
+                    chain(sub_directories, files),
                 )
-            ):
-                shutil.rmtree(directory)
-                number_of_deleted_directories += 1
-    if number_of_deleted_directories:
+            )
+        ):
+            if dry_run:
+                print(f"rm -R {sub_directory}")
+            else:
+                shutil.rmtree(sub_directory)
+            number_of_deleted_directories += 1
+    if number_of_deleted_directories and not dry_run:
         number_of_deleted_directories += delete_empty_directories(
-            root_directory,
-            exclude_directories=exclude_directories,
-            _recurrence=False,
+            directory,
+            exclude=exclude,
+            dry_run=dry_run,
+            _recurrence=True,
         )
-    if _recurrence and number_of_deleted_directories:
-        print(f"Deleted {number_of_deleted_directories} empty directories")
+    if (not _recurrence) and number_of_deleted_directories:
+        if not dry_run:
+            print(f"Deleted {number_of_deleted_directories} empty directories")
     return number_of_deleted_directories
 
 
 def delete_ignored(
-    root_directory: str = ".",
-    exclude_directories: FrozenSet[str] = frozenset(),
+    directory: str = ".",
+    exclude: FrozenSet[str] = frozenset(),
+    dry_run: bool = False,
 ) -> None:
     """
     Delete files which are ignored by Git.
@@ -150,18 +212,46 @@ def delete_ignored(
     Parameters:
 
     - root_directory (str): The root project directory.
-    - exclude_directories ({str}) = frozenset(): A `frozenset` of directories
-      which should not be touched.
+    - exclude ({{str}}) = {EXCLUDE_DIRECTORIES}: A `frozenset` of
+      directories to leave untouched.
     """
-    for path in get_ignored_files(
-        directory=root_directory, exclude_directories=exclude_directories
-    ):
-        os.remove(path)
+    sub_directory_name: str
+    directories_files: Dict[str, Set[str]] = {}
+    paths: Set[str]
+    path: str
+    for path in get_ignored_files(directory=directory, exclude=exclude):
+        sub_directory_name = ""
+        if "/" in path:
+            sub_directory_name = path.split("/")[0]
+        if sub_directory_name not in directories_files:
+            directories_files[sub_directory_name] = set()
+        directories_files[sub_directory_name].add(
+            os.path.join(directory, path)
+        )
+    for sub_directory_name, paths in directories_files.items():
+        number_of_files: int = len(paths)
+        if not dry_run:
+            print(
+                f"Deleting {number_of_files} ignored "
+                f'file{"s" if number_of_files > 1 else ""} in '
+                f"{sub_directory_name}"
+            )
+        for path in paths:
+            if dry_run:
+                print(f"rm {path}")
+            else:
+                os.remove(path)
+
+
+delete_ignored.__doc__ = delete_ignored.__doc__.format(  # type: ignore
+    EXCLUDE_DIRECTORIES=repr(frozenset(DEFAULT_EXCLUDE))
+)
 
 
 def clean(
-    root_directory: str = ".",
-    exclude_directories: FrozenSet[str] = EXCLUDE_DIRECTORIES,
+    directory: str = ".",
+    exclude: FrozenSet[str] = frozenset(DEFAULT_EXCLUDE),
+    dry_run: bool = False,
 ) -> None:
     """
     Cleanup (delete) files which are ignored by Git and subsequently delete all
@@ -170,33 +260,51 @@ def clean(
     Parameters:
 
     - root_directory (str) = ".": The project's root directory.
-    - exclude_directories ({{str}}) = {EXCLUDE_DIRECTORIES}: A `frozenset` of
+    - exclude ({{str}}) = {EXCLUDE_DIRECTORIES}: A `frozenset` of
       directories to leave untouched.
     """
-    delete_ignored(root_directory, exclude_directories=exclude_directories)
-    delete_empty_directories(
-        root_directory, exclude_directories=exclude_directories
-    )
+    delete_ignored(directory, exclude=exclude, dry_run=dry_run)
+    delete_empty_directories(directory, exclude=exclude, dry_run=dry_run)
 
 
 clean.__doc__ = clean.__doc__.format(  # type: ignore
-    EXCLUDE_DIRECTORIES=EXCLUDE_DIRECTORIES
+    EXCLUDE_DIRECTORIES=repr(frozenset(DEFAULT_EXCLUDE))
 )
 
 
 def main() -> None:
     parser: argparse.ArgumentParser = argparse.ArgumentParser(
-        prog="daves-dev-tool clean"
+        prog="daves-dev-tool clean",
+        description=(
+            "This command removes all files from your project directory which "
+            "are ignored by git, unless matching one of the EXCLUDE glob "
+            "patterns."
+        ),
+    )
+    exclude: str
+    default_exclude_parameters: str = " ".join(
+        f"-e {exclude}" for exclude in DEFAULT_EXCLUDE
     )
     parser.add_argument(
-        "--exclude",
         "-e",
-        action="store",
+        "--exclude",
         type=str,
-        default=None,
+        action="append",
         help=(
-            'A path list of sub-directories to exclude (separated by ":" or '
-            '";", depending on the operating system).'
+            "One or more glob patterns indicating files/directories to "
+            "exclude from cleanup. The default values are: `"
+            f"{default_exclude_parameters}`."
+        ),
+    )
+    parser.add_argument(
+        "-dr",
+        "--dry-run",
+        default=False,
+        action="store_const",
+        const=True,
+        help=(
+            "Instead of executing the cleanup, just print the shell "
+            "commands (a list of `rm FILE` commands)."
         ),
     )
     parser.add_argument(
@@ -208,12 +316,11 @@ def main() -> None:
     )
     arguments: argparse.Namespace = parser.parse_args()
     clean(
-        root_directory=arguments.directory,
-        exclude_directories=(
-            frozenset(os.path.split(arguments.exclude))
-            if arguments.exclude
-            else EXCLUDE_DIRECTORIES
+        directory=arguments.directory,
+        exclude=frozenset(
+            arguments.exclude if arguments.exclude else DEFAULT_EXCLUDE
         ),
+        dry_run=arguments.dry_run,
     )
